Move star from modules to components, fix _DMS not defined
